update to m.py
