--- README
@@ -1,3 +1,0 @@
-Kode yang berada di branch main itu kode backend murni ya wa
-Kode yang nantinya bakal dicustom UI disimpan di branch fork
-Jadi kamu pull sama push langsung ke branch fork
